Move reflink to os dependent file. OSX docker client fully functionnal.
